Partial work on quest status markers over quest givers. Will work on after I get some sleep. Brain is fried.

+++ src/server/game/Entities/Creature/GossipDef.h
@@ -271,7 +271,7 @@ class PlayerMenu
         /*********************************************************/
         /***                    QUEST SYSTEM                   ***/
         /*********************************************************/
-        void SendQuestGiverStatus(uint8 questStatus, uint64 npcGUID);
+        void SendQuestGiverStatus(uint32 questStatus, uint64 npcGUID);
 
         void SendQuestGiverQuestList(QEmote eEmote, const std::string& Title, uint64 npcGUID);
